chore: move ResourceNotFoundException to internal
